Move mosq functions to new file: Mosquito.cpp
